Added chatCommand function, base for many future uses.
It seperates elements in a string by spaces, then puts them into a
string vector for use, ment to return false if no command is found,
return true if command is found.

Should also put an error in the chat if there's not enough arguments for
a specific command, or some such.