Working on setting up Bresenham's line for vision
It's not going well, Turns out it swaps the numbers to start from the
lower position(Value wise), So it starts ON the target in half the
cases, making it useless.

It'll require checking if it hits the user AND the target to be useful,
Not sure it's worth it.